+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   - Breaking. Some changes to make movingai-rust compatible to Rust API guidelines.
     - `get_height` renamed into `height`, `get_width` renamed into `width`, `get_cell` renamed into `get`.
     - Rename `coords_iter` into `coords` and `Map2DCoordsIter` into `CoordsIter`.
+  - Add `Debug`, `PartialEq` and `Clone` to `SceneRecord`.
 
 ## [0.5.1] - 2018-02-07